RichGirls/requirements.txt
django==4.2.20
djangorestframework==3.16.0
django-cors-headers==4.7.0
numpy>=1.21,<2.0
pandas>=2.2.0
networkx==3.2.1
plotly==6.0.1
seaborn==0.13.2
matplotlib>=3.8.0
scipy>=1.13.0 